organize page components
