+++ types/image.ts
@@ -1,0 +1,1 @@
+export type Img = { id: string; url: string };
